% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two_numbers.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Add Two Numbers}
\usage{
add_two_numbers(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{Another number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add_two_numbers(1, 2)
}
