% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_two_chars.R
\name{last_two_chars}
\alias{last_two_chars}
\title{Last Two Characters}
\usage{
last_two_chars(word)
}
\arguments{
\item{word}{A character string.}
}
\value{
The last two characters of \code{word}.
}
\description{
This function returns the last two characters of a given string.
}
\examples{
last_two_chars("hello")
}
